Add functional tests for DemoController
A new set of functional tests have been added specifically for the DemoController, along with the necessary configuration files. Additionally, an update was done to the composer.json file to include the Laminas Router Attributes package in the "require-dev" and "suggest" sections, which is essential for the creation and suggestion of Symfony route configurations.